packages/crawler/pyproject.toml
[project]
name = "crawler"
version = "0.1.0"
description = "Crawler"
readme = "README.md"
authors = [
    { name = "Lefteris", email = "eldelielev@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "sqlalchemy>=2.0.38",
    "strawberry-graphql[fastapi]>=0.255.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/crawler"]
